Setup reactive image component
We want to allow the browser to select the image format and the size
that is most appropriate to do so. We use srcset, we took inspiration
from how medium set this up.

By looking at what was going on in this article:

https://medium.com/@themaheshvaishnav/a-beginners-guide-to-responsive-images-in-html-9c5df5a5146d#:~:text=Using%20the%20srcset,the%20size%20of%20the%20device.&text=You%20can%20use%20srcset%20to,fit%20for%20that%20image%20size.

I could never get media queries working.